// front-end-app/quanly-thuysan/src/app/model/phanphoi.ts
export class PhanPhoi {
  $class: string;
  idgdpp: string;
  sanphamnoi: string;
  sanphamngoai: string;
  nguyenlieu: string;
  cspp: string;
  soluongnhan: number;
  tgnhan: string;
  masolonhan: string;
  tennguoinhan: string;
  phanloai: string;
  diadiemnhan: string;
  masanpham: string;
  soluonggiao: number;
  tggiao: string;
  masologiao: string;
  tennguoigiao: string;
  phuongtien: string;
  csbl: string;
  nhaphanphoi: string;
}